refactor: Introduce units.datetime module
