+++ pyaerocom/colocation/colocation_utils.py
@@ -26,14 +26,14 @@ from pyaerocom.exceptions import (
 )
 from pyaerocom.filter import Filter
 from pyaerocom.griddeddata import GriddedData
-from pyaerocom.units.helpers import get_lowest_resolution
+from pyaerocom.units.datetime import get_lowest_resolution
 from pyaerocom.helpers import (
     isnumeric,
     make_datetime_index,
     to_pandas_timestamp,
 )
 from pyaerocom.time_resampler import TimeResampler
-from pyaerocom.units.tstype import TsType
+from pyaerocom.units.datetime import TsType
 
 if TYPE_CHECKING:
     from pyaerocom.ungriddeddata import UngriddedData
