+++ pyaerocom/combine_vardata_ungridded.py
@@ -3,7 +3,7 @@ import numpy as np
 from pyaerocom._lowlevel_helpers import invalid_input_err_str
 from pyaerocom.colocation.colocation_utils import _colocate_site_data_helper
 from pyaerocom.geodesy import find_coord_indices_within_distance
-from pyaerocom.units.helpers import sort_ts_types
+from pyaerocom.units.datetime import sort_ts_types
 from pyaerocom.obs_io import ObsVarCombi
 from pyaerocom.stationdata import StationData
