+++ tests/units/test_helpers.py
@@ -2,34 +2,13 @@ from datetime import timedelta
 import numpy as np
 import pytest
 from pyaerocom.units.helpers import (
-    get_lowest_resolution,
     seconds_in_periods,
-    sort_ts_types,
     get_standard_unit,
 )
 
 
-@pytest.mark.parametrize(
-    "input,expected",
-    (
-        pytest.param(
-            ["monthly", "weekly", "daily", "hourly"], ["hourly", "daily", "weekly", "monthly"]
-        ),
-        pytest.param(
-            ["3daily", "4daily", "6daily", "13daily"], ["3daily", "4daily", "6daily", "13daily"]
-        ),
-    ),
-)
-def test_sort_tstypes(input, expected):
-    assert sort_ts_types(input) == expected
-
-
 def test_get_standard_unit():
     assert get_standard_unit("ec550aer") == "1/km"
-
-
-def test_get_lowest_resolution():
-    assert get_lowest_resolution("3hourly", "hourly", "monthly", "yearly") == "yearly"
 
 
 @pytest.mark.parametrize(
